requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bitarray==3.6.0
certifi==2025.4.26
charset-normalizer==3.4.2
ckzg==2.1.1
cytoolz==1.0.1
deprecation==2.1.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
frozenlist==1.6.0
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hexbytes==1.3.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
multidict==6.4.4
packaging==25.0
parsimonious==0.10.0
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.1
psycopg2-binary==2.9.10
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.4.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyunormalize==16.0.0
realtime==2.4.3
regex==2025.7.34
requests==2.32.3
rlp==4.1.0
six==1.17.0
sniffio==1.3.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.2
supafunc==0.9.4
toolz==1.0.0
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
web3==7.13.0
websockets==14.2
yarl==1.20.0
fal-client==0.7.0

